# lab-5/src/main/kotlin/org/nechaev/pomodoro/api/TimersApi.kt
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.12.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
*/
package org.nechaev.pomodoro.api

import org.nechaev.pomodoro.model.CreateTimerRequest
import org.nechaev.pomodoro.model.ErrorResponse
import org.nechaev.pomodoro.model.Timer
import io.swagger.v3.oas.annotations.*
import io.swagger.v3.oas.annotations.enums.*
import io.swagger.v3.oas.annotations.media.*
import io.swagger.v3.oas.annotations.responses.*
import io.swagger.v3.oas.annotations.security.*
import org.springframework.http.HttpStatus
import org.springframework.http.MediaType
import org.springframework.http.ResponseEntity

import org.springframework.web.bind.annotation.*
import org.springframework.validation.annotation.Validated
import org.springframework.web.context.request.NativeWebRequest
import org.springframework.beans.factory.annotation.Autowired

import jakarta.validation.constraints.DecimalMax
import jakarta.validation.constraints.DecimalMin
import jakarta.validation.constraints.Email
import jakarta.validation.constraints.Max
import jakarta.validation.constraints.Min
import jakarta.validation.constraints.NotNull
import jakarta.validation.constraints.Pattern
import jakarta.validation.constraints.Size
import jakarta.validation.Valid

import kotlin.collections.List
import kotlin.collections.Map

@RestController
@Validated
interface TimersApi {

    fun getDelegate(): TimersApiDelegate = object: TimersApiDelegate {}

    @Operation(
        tags = ["timers",],
        summary = "Завершение таймера",
        operationId = "completeTimer",
        description = """Переводит таймер в состояние COMPLETED. После завершения таймер
не может быть запущен повторно. Допускается из любого состояния, кроме COMPLETED.
""",
        responses = [
            ApiResponse(responseCode = "200", description = "Таймер завершён", content = [Content(schema = Schema(implementation = Timer::class))]),
            ApiResponse(responseCode = "404", description = "Таймер с указанным ID не найден", content = [Content(schema = Schema(implementation = ErrorResponse::class))]),
            ApiResponse(responseCode = "409", description = "Операция невозможна в текущем состоянии таймера", content = [Content(schema = Schema(implementation = ErrorResponse::class))])
        ]
    )
    @RequestMapping(
            method = [RequestMethod.POST],
            value = ["/timers/{id}/complete"],
            produces = ["application/json"]
    )
    fun completeTimer(@Parameter(description = "Уникальный идентификатор таймера", required = true) @PathVariable("id") id: kotlin.Long): ResponseEntity<Timer> {
        return getDelegate().completeTimer(id)
    }

    @Operation(
        tags = ["timers",],
        summary = "Создание нового таймера",
        operationId = "createTimer",
        description = """Создаёт таймер в состоянии CREATED. По умолчанию длительность - 25 минут.
Таймер не запускается автоматически, для старта нужно вызвать `/timers/{id}/start`.
""",
        responses = [
            ApiResponse(responseCode = "201", description = "Таймер успешно создан", content = [Content(schema = Schema(implementation = Timer::class))]),
            ApiResponse(responseCode = "400", description = "Некорректный запрос (например, не указано имя)", content = [Content(schema = Schema(implementation = ErrorResponse::class))])
        ]
    )
    @RequestMapping(
            method = [RequestMethod.POST],
            value = ["/timers"],
            produces = ["application/json"],
            consumes = ["application/json"]
    )
    fun createTimer(@Parameter(description = "", required = true) @Valid @RequestBody createTimerRequest: CreateTimerRequest): ResponseEntity<Timer> {
        return getDelegate().createTimer(createTimerRequest)
    }

    @Operation(
        tags = ["timers",],
        summary = "Получение списка всех таймеров",
        operationId = "getAllTimers",
        description = """Возвращает все созданные таймеры с актуальным значением оставшегося времени.""",
        responses = [
            ApiResponse(responseCode = "200", description = "Список таймеров", content = [Content(array = ArraySchema(schema = Schema(implementation = Timer::class)))])
        ]
    )
    @RequestMapping(
            method = [RequestMethod.GET],
            value = ["/timers"],
            produces = ["application/json"]
    )
    fun getAllTimers(): ResponseEntity<List<Timer>> {
        return getDelegate().getAllTimers()
    }

    @Operation(
        tags = ["timers",],
        summary = "Получение таймера по идентификатору",
        operationId = "getTimerById",
        description = """Возвращает таймер с актуальным значением `remainingSeconds`.
Для запущенного таймера оставшееся время пересчитывается при каждом запросе.
""",
        responses = [
            ApiResponse(responseCode = "200", description = "Таймер найден", content = [Content(schema = Schema(implementation = Timer::class))]),
            ApiResponse(responseCode = "404", description = "Таймер с указанным ID не найден", content = [Content(schema = Schema(implementation = ErrorResponse::class))])
        ]
    )
    @RequestMapping(
            method = [RequestMethod.GET],
            value = ["/timers/{id}"],
            produces = ["application/json"]
    )
    fun getTimerById(@Parameter(description = "Уникальный идентификатор таймера", required = true) @PathVariable("id") id: kotlin.Long): ResponseEntity<Timer> {
        return getDelegate().getTimerById(id)
    }

    @Operation(
        tags = ["timers",],
        summary = "Запуск таймера",
        operationId = "startTimer",
        description = """Переводит таймер в состояние RUNNING и начинает отсчёт времени.
Допускается из состояний CREATED и PAUSED. При возобновлении после паузы
отсчёт продолжается с того места, где был остановлен.
""",
        responses = [
            ApiResponse(responseCode = "200", description = "Таймер запущен", content = [Content(schema = Schema(implementation = Timer::class))]),
            ApiResponse(responseCode = "404", description = "Таймер с указанным ID не найден", content = [Content(schema = Schema(implementation = ErrorResponse::class))]),
            ApiResponse(responseCode = "409", description = "Операция невозможна в текущем состоянии таймера", content = [Content(schema = Schema(implementation = ErrorResponse::class))])
        ]
    )
    @RequestMapping(
            method = [RequestMethod.POST],
            value = ["/timers/{id}/start"],
            produces = ["application/json"]
    )
    fun startTimer(@Parameter(description = "Уникальный идентификатор таймера", required = true) @PathVariable("id") id: kotlin.Long): ResponseEntity<Timer> {
        return getDelegate().startTimer(id)
    }

    @Operation(
        tags = ["timers",],
        summary = "Приостановка таймера",
        operationId = "stopTimer",
        description = """Переводит таймер в состояние PAUSED. Накопленное время текущей сессии
сохраняется, и при следующем запуске отсчёт продолжится с того же значения.
Допускается только из состояния RUNNING.
""",
        responses = [
            ApiResponse(responseCode = "200", description = "Таймер приостановлен", content = [Content(schema = Schema(implementation = Timer::class))]),
            ApiResponse(responseCode = "404", description = "Таймер с указанным ID не найден", content = [Content(schema = Schema(implementation = ErrorResponse::class))]),
            ApiResponse(responseCode = "409", description = "Операция невозможна в текущем состоянии таймера", content = [Content(schema = Schema(implementation = ErrorResponse::class))])
        ]
    )
    @RequestMapping(
            method = [RequestMethod.POST],
            value = ["/timers/{id}/stop"],
            produces = ["application/json"]
    )
    fun stopTimer(@Parameter(description = "Уникальный идентификатор таймера", required = true) @PathVariable("id") id: kotlin.Long): ResponseEntity<Timer> {
        return getDelegate().stopTimer(id)
    }
}